requirements.txt
fastapi==0.108.0
xgboost==2.1.1
pandas==2.2.0
numpy==1.26.3
uvicorn==0.25.0
joblib==1.3.2
scikit-learn==1.4.0
textblob==0.17.1
pyspellchecker==0.8.2